requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
email-validator==2.1.0
psycopg==3.1.18
injective-py>=1.10.0
ecdsa==0.18.0
web3>=7.0.0
protobuf>=5.26.1
